requirements.txt
anyio
click
httpx
mcp
python-dotenv
jira
pydantic
uvicorn
starlette 